Rename target files: main.cpp and board.h
